Extract Tiff decoder useDecoderTiff from body index
