+++ src/hooks/index.tsx
@@ -1,0 +1,3 @@
+import useDecoderTiff from "./useDecoderTiff";
+
+export { useDecoderTiff };
